requirements.txt
dvc
dvc[gdrive]
sklearn
pandas
pytest
tox
flake8
flask
gunicorn
cassandra-driver
